adds post method for postgres model
