+++ src/api/routes/route-handlers/create-book.js
@@ -1,40 +1,17 @@
 'use strict';
 
-const client = require('../../../../index.js');
-
-
-function createShelf(shelf) {
-  let normalizedShelf = shelf.toLowerCase();
-  let SQL1 = `SELECT id from bookshelves where name=$1;`;
-  let values1 = [normalizedShelf];
-
-  return client.query(SQL1, values1)
-    .then(results => {
-      if (results.rowCount) {
-        return results.rows[0].id;
-      } else {
-        let INSERT = `INSERT INTO bookshelves(name) VALUES($1) RETURNING id;`;
-        let insertValues = [shelf];
-
-        return client.query(INSERT, insertValues)
-          .then(results => {
-            return results.rows[0].id;
-          })
-      }
-    })
-}
-
 module.exports = (request, response, next) => {
-  console.log('create book');
+  console.log('create-book.js');
 
-  console.log(request.body.bookshelf);
-  createShelf(request.body.bookshelf)
+  // returns the unique identifier for a bookshelf
+  request.model.createShelf(request.body.bookshelf)
     .then(id => {
       let { title, author, isbn, image_url, description } = request.body;
-      let SQL = 'INSERT INTO books(title, author, isbn, image_url, description, bookshelf_id) VALUES($1, $2, $3, $4, $5, $6) RETURNING id;';
       let values = [title, author, isbn, image_url, description, id];
 
-      client.query(SQL, values)
+      // adds a book to the database with values from the request.body
+      request.model.post(values)
+      // client.query(SQL, values)
         .then(result => response.redirect(`/books/${result.rows[0].id}`))
         .catch(error => {
           response.render('pages/error', { error: error })
