+++ src/api/routes/routes.js
@@ -35,7 +35,7 @@ router.use(modelFinder);
 router.get('/', getBooks); // reads from DB //
 router.post('/searches', createSearch); // superagent
 router.get('/searches/new', newSearch); // renders a page
-router.get('/books/:id', getBook); // reads 1 from DB
+router.get('/books/:id', getBook); // reads 1 from DB //ß
 router.post('/books', createBook); // adds 1 to DB
 router.put('/books/:id', updateBook); // edits 1 to DB
 router.delete('/books/:id', deleteBook); // deletes 1 from DB
